# backend/swagger/api-docs.yml
openapi: 3.0.0
info:
  title: ZeYang API
  description: ZeYang 房地產專案管理系統 API 文檔
  version: 2.1.0
  contact:
    name: API Support
    email: dev@ZeYang.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/api
    description: Development server
  - url: https://api.ZeYang.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: object
          properties:
            code:
              type: string
              example: ERROR_CODE
            details:
              type: string
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
...
    Project:
      type: object
      properties:
        id:
          type: integer
          description: 專案 ID
        identifier:
          type: string
          description: 專案識別碼
        name:
          type: string
          description: 專案名稱
        nameEn:
          type: string
          description: 英文名稱
        type:
          type: string
          enum: [residential, commercial, mixed, other]
          description: 專案類型
        status:
          type: string
          enum: [planning, in_progress, completed, on_hold]
          description: 專案狀態
        description:
          type: string
          description: 專案描述
        descriptionEn:
          type: string
          description: 英文描述
        location:
          type: string
          description: 位置
        locationEn:
          type: string
          description: 英文位置
        price:
          type: number
          description: 價格
        priceMin:
          type: number
          description: 最低價格
        priceMax:
          type: number
          description: 最高價格
        area:
          type: number
          description: 面積
        areaMin:
          type: number
          description: 最小面積
        areaMax:
          type: number
          description: 最大面積
        developer:
          type: string
          description: 開發商
        architect:
          type: string
          description: 建築師
        yearStarted:
          type: integer
          description: 開工年份
        yearCompleted:
          type: integer
          description: 完工年份
        units:
          type: integer
          description: 戶數
        floors:
          type: integer
          description: 樓層數
        features:
          type: object
          description: 特色功能
        featuresEn:
          type: object
          description: 英文特色功能
        mainImage:
          type: string
          description: 主圖片 URL
        videoUrl:
          type: string
          description: 影片 URL
        website:
          type: string
          description: 官網 URL
        brochureUrl:
          type: string
          description: 手冊 URL
        isFeatured:
          type: boolean
          description: 是否為精選
        displayOrder:
          type: integer
          description: 顯示順序
        viewCount:
          type: integer
          description: 瀏覽次數
        created_at:
          type: string
          format: date-time
          description: 建立時間
        updated_at:
          type: string
          format: date-time
          description: 更新時間
...
    ProjectImage:
      type: object
      properties:
        id:
          type: integer
          description: 圖片 ID
        project_id:
          type: integer
          description: 專案 ID
        filename:
          type: string
          description: 檔案名稱
        originalName:
          type: string
          description: 原始檔案名稱
        path:
          type: string
          description: 檔案路徑
        url:
          type: string
          description: 圖片 URL
        thumbnailUrl:
          type: string
          description: 縮圖 URL
        alt:
          type: string
          description: 替代文字
        caption:
          type: string
          description: 圖片說明
        isMain:
          type: boolean
          description: 是否為主圖
        displayOrder:
          type: integer
          description: 顯示順序
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 總筆數
        page:
          type: integer
          description: 目前頁數
        limit:
          type: integer
          description: 每頁筆數
        pages:
          type: integer
          description: 總頁數
        hasNext:
          type: boolean
          description: 是否有下一頁
        hasPrev:
          type: boolean
          description: 是否有上一頁
paths:
...
  /projects:
    get:
      tags:
        - Projects
      summary: 取得專案列表
      description: 取得專案列表，支援分頁、篩選和排序
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 頁碼
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每頁筆數
        - in: query
          name: type
          schema:
            type: string
            enum: [residential, commercial, mixed, other]
          description: 專案類型
        - in: query
          name: status
          schema:
            type: string
            enum: [planning, in_progress, completed, on_hold]
          description: 專案狀態
        - in: query
          name: isFeatured
          schema:
            type: boolean
          description: 是否為精選專案
        - in: query
          name: search
          schema:
            type: string
          description: 搜尋關鍵字
        - in: query
          name: orderBy
          schema:
            type: string
            enum: [displayOrder, created_at, updated_at, viewCount, name]
            default: displayOrder
          description: 排序欄位
        - in: query
          name: orderDir
          schema:
            type: string
            enum: [ASC, DESC]
            default: ASC
          description: 排序方向
      responses:
        200:
          description: 專案列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Project'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
    
    post:
      tags:
        - Projects
      summary: 建立新專案
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - category
                - location
              properties:
                title:
                  type: string
                  description: 專案標題
                category:
                  type: string
                  enum: [住宅, 商辦, 公共工程, 其他]
                  description: 專案分類
                location:
                  type: string
                  description: 專案位置
                status:
                  type: string
                  enum: [planning, pre_sale, on_sale, sold_out, completed]
                year:
                  type: integer
                  minimum: 1900
                  maximum: 2100
                is_featured:
                  type: boolean
                tags:
                  type: array
                  items:
                    type: string
                features:
                  type: array
                  items:
                    type: string
      responses:
        201:
          description: 專案建立成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        422:
          $ref: '#/components/responses/ValidationError'
...
  /projects/{identifier}:
    get:
      tags:
        - Projects
      summary: 取得單一專案
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
          description: 專案識別碼或 UUID
      responses:
        200:
          description: 專案詳細資訊
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          project:
                            $ref: '#/components/schemas/Project'
        404:
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Projects
      summary: 更新專案
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                category:
                  type: string
                  enum: [住宅, 商辦, 公共工程, 其他]
                status:
                  type: string
                  enum: [planning, pre_sale, on_sale, sold_out, completed]
                year:
                  type: integer
                is_featured:
                  type: boolean
                tags:
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: 專案更新成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Projects
      summary: 刪除專案
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
        - in: query
          name: hard
          schema:
            type: boolean
            default: false
          description: 是否永久刪除
      responses:
        200:
          description: 專案刪除成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
...
  /projects/reorder:
    put:
      tags:
        - Projects
      summary: 批量更新專案顯示順序
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orders
              properties:
                orders:
                  type: array
                  items:
                    type: object
                    required:
                      - identifier
                      - displayOrder
                    properties:
                      identifier:
                        type: string
                      displayOrder:
                        type: integer
                        minimum: 0
      responses:
        200:
          description: 順序更新成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  # 標籤管理端點
...
  /projects/{identifier}/images:
    get:
      tags:
        - Project Images
      summary: 取得專案圖片列表
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
          description: 專案識別碼
        - in: query
          name: image_type
          schema:
            type: string
            enum: [gallery, hero, thumbnail, floor_plan, other]
          description: 圖片類型
      responses:
        200:
          description: 專案圖片列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          images:
                            type: array
                            items:
                              $ref: '#/components/schemas/ProjectImage'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
    
    post:
      tags:
        - Project Images
      summary: 上傳專案圖片
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                image_type:
                  type: string
                  enum: [gallery, hero, thumbnail, floor_plan, other]
                  default: gallery
                alt_text:
                  type: string
      responses:
        201:
          description: 圖片上傳成功
        400:
          description: 沒有檔案上傳
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  /projects/{identifier}/images/{imageId}:
    put:
      tags:
        - Project Images
      summary: 更新圖片詳情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alt_text:
                  type: string
                display_order:
                  type: integer
                  minimum: 0
                image_type:
                  type: string
                  enum: [gallery, hero, thumbnail, floor_plan, other]
      responses:
        200:
          description: 圖片更新成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Project Images
      summary: 刪除圖片
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 圖片刪除成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
  /projects/{identifier}/images/{imageId}/set-main:
    post:
      tags:
        - Project Images
      summary: 設為主圖片
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 主圖片設定成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  /projects/{identifier}/images/reorder:
    put:
      tags:
        - Project Images
      summary: 重新排序圖片
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orders
              properties:
                orders:
                  type: array
                  items:
                    type: object
                    required:
                      - id
                      - display_order
                    properties:
                      id:
                        type: integer
                      display_order:
                        type: integer
                        minimum: 0
      responses:
        200:
          description: 圖片順序更新成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  /projects/{identifier}/images/bulk-delete:
    post:
      tags:
        - Project Images
      summary: 批量刪除圖片
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - imageIds
              properties:
                imageIds:
                  type: array
                  items:
                    type: integer
      responses:
        200:
          description: 圖片批量刪除成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'

  # 統計相關端點
...
  /statistics/projects:
    get:
      tags:
        - Statistics
      summary: 取得專案統計
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
      responses:
        200:
          description: 專案統計資料
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
...
components:
  responses:
    Unauthorized:
      description: 未授權
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 權限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 資源未找到
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: 驗證錯誤
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
...